Reorganizing bls numba utils
